tdf#130857 tdf#168646 qt weld: Implement showing extended tips
Extended tips currently use the text set as the accessible
description in .ui files and is shown for the tooltip if
extended tips are enabled in the options.

Implement logic to show these for the case of native Qt widgets
being used, i.e. for qt6 with SAL_VCL_QT_USE_WELDED_WIDGETS=1.

Sample scenario where this is used:

* at "Tools" -> "Options" -> "LibreOfficeDev" -> "General",
  enable "Extended Tips" and close dialog using "OK"
* open spell check dialog using "Tools" -> "Options" -> "Spelling"
* click "Options" button
* in the "Options" dailog, click the "New" button to open the
  "New Dictionary" dialog
* hover over the line edit labelled "Name"

-> The tooltip "Specifies the name of the new custom dictionary"
shows up.

Change-Id: I0aeba9a35a1f30cb5c0c50eec75e37e3a8eb1d17
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191799
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins