tdf#130857 tdf#168646 qt weld: Call base class tooltip logic
Refactor, deduplicate, and refine the custom logic for
tooltip handling for QtInstanceWidget and subclasses:

First, deduplicate and refactor, by making
QtInstanceWidget::handleToolTipEvent virtual and
overriding it in QtInstanceDrawingArea and
QtInstanceIconView. This also means that their
eventFilter overrides no longer need to handle the QToolTip
QEvent::ToolTip case directly anymore, as the
QtInstanceWidget implementation now calls their
QtInstanceWidget::handleToolTipEvent overrides.

Rename QtInstanceTreeView::handleToolTipEvent
to QtInstanceTreeView::handleViewPortToolTipEvent
as that one is different: It handles the tooltip
event for its QTreeView's viewPort(), not for the
QTreeView widget itself.

Second, in QtInstanceIconView::handleToolTipEvent
and QtInstanceDrawingArea::handleToolTipEvent,
call the base class implementation when there is
no tooltip according to the IconView or DrawingArea
specific logic, to ensure that the extended tip logic
implemented in QtInstanceWidget::handleToolTipEvent
gets triggered.

The latter is similar to what

    Change-Id: I2918cc3377ec53e649d54cceaf0c9cb53effe58f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 2 23:17:33 2025 +0200

        tdf#168646 vcl: Propagate undhandled tooltip event to base class

implemented for the SalInstanceDrawing case, i.e. the vcl
immplementation.

This makes the extended tip for the case of the "Dimension Line"
dialog described in tdf#168646 work with qt6 and
SAL_VCL_QT_USE_WELDED_WIDGETS=1 in a WIP branch
where support for using native Qt widgets for that dialog
is declared.

Change-Id: I00cd0486911dd65d59315d02c1810ccdcc763d0d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191803
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>