bin/find-unneeded-includes: don't trip on BOX2D_HEADER in --sanitycheck
This is tricky because a header name is defined in a macro
and I had an exclusion for this external libs wrapper header.

But --sanitycheck does not see that, as it looks for presence of
exclusion list items.

Move the exclusion into the script, so that find-unneeded-includes
ignores it when checking the IWYU output and --sanitycheck does not
have to deal with this special rare case

Change-Id: I55d61630dd63a48d9b264df4b714dad999d1eacd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191827
Reviewed-by: Gabor Kelemen <gabor.kelemen.extern@allotropia.de>
Tested-by: Jenkins