tdf#168662 sw markdown export: handle anchored images
The bugdoc has a to-char anchored image, exporting that to markdown the
image is lost.

This happens because anchored images are not represented in the doc
model as a SwFormatFlyCnt on a dummy character.

Fix the problem similar to what the HTML export does with its
SwHTMLWriter::OutFlyFrame() code:
1) Collect the position of all fly formats at the start
2) In OutMarkdown_SwTextNode(), take the flys anchored to a given text
   node out of that container
3) In OutFormattingChange(), emit the anchored images using the markdown
   image markup, which will turn them into inline images

Draw shapes are ignored for now, since the current fly export code takes
the object title / description directly from the fly frame format.

Change-Id: I8061a4ea51b2c24a7d48daca0b3c62d79d3e7a3e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191811
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>