tdf#168540 Restore automatically enabling RTL/CTL based on locale
This change makes LibreOffice automatically enable RTL/CTL features
based on locale. On Windows, in addition to locale, RTL/CTL features are
enabled if RTL/CTL is installed as an input language.

The code to do this was originally deleted roughly a decade ago, as part
of an effort to make RTL/CTL support enabled by default. That default
doesn't seem to be effective anymore, and without the removed code, this
left LO in a state where all RTL/CTL users had to manually enable
RTL/CTL support in order to see the features they expect.

This reverts commit 23b92427e3c77fa8d04679ddcaf2dbd3e3580c5d.

Change-Id: I15a9bf9ebdb759b49743d464a2ce6f9305e4c8d6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191835
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>