tdf#130857 qt weld: Evaluate GtkTextView::editable
Evaluate the GtkTextView::editable [1] property.
This ensures that the QPlainTextEdit used in
QtInstanceTextView is set to readonly when
GtkTextView::editable is set to the (non-default)
value of false.

This will e.g. be relevant for the mail merge
test settings dialog that can be triggered as
follows:

* "Tools" -> "Options" -> "LibreOfficeDev Writer"
  -> "Mail Merge Email"
* click the "Test Settings" button

(Support for using native Qt widgets isn't declared
for that dialog yet. This is one step towards making
that possible.)

[1] https://docs.gtk.org/gtk3/property.TextView.editable.html

Change-Id: I2436892589b6ba521ecdfa262a46046c7b6ea2d2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191867
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>