tdf#130857 qt weld: Evaluate "GtkTextView" text
A GtkTextView's text is set in a GtkTextBuffer
that is specified via the GtkTextView::buffer property [1].

Evaluate that and the GtkTextBuffer::text property [2]
to set the initial text for "GtkTextBuffer" objects
as specified in the .ui file.

One example where such text is specified is in
the mail merge test settings dialog that can be triggered
as follows:

* "Tools" -> "Options" -> "LibreOfficeDev Writer"
  -> "Mail Merge Email"
* click the "Test Settings" button

(Support for using native Qt widgets isn't declared
for that dialog yet. This is one step towards making
that possible in the future.)

[1] https://docs.gtk.org/gtk3/property.TextView.buffer.html
[2] https://docs.gtk.org/gtk3/property.TextBuffer.text.html

Change-Id: I174380a021d859366a33eb96d96bfe9a52d452e2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191868
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins