tdf#130857 Make mail merge test account settings dialog async
In its ctor, SwTestAccountSettingsDialog posts an event
that is responsible for actually running the account
settings test:

    m_pPostedEvent = Application::PostUserEvent(LINK(this, SwTestAccountSettingsDialog, TestHdl));

For the case of a native Qt dialog (QtInstanceDialog),
that event would only be processed after the dialog
gets closed when the dialog is run synchronously,
so no test result would be shown in the dialog at all
(once support for it is declared for using native Qt widgets,
which will be done in an upcoming commit).

Run the dialog asynchronously to make sure events
get processed while it is being run. Explicitly
make it modal to make sure it remains on top of
the parent.

The dialog can be triggered as follows:

* "Tools" -> "Options" -> "LibreOfficeDev Writer"
  -> "Mail Merge Email"
* click the "Test Settings" button

Change-Id: I0ffb0c0a59690917217fd3d5d7cdf8dfb092e364
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191869
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>