tdf#165785: basctl: Activate Core UI Interactivity and Member View
This patch brings the Object Browser's user interface to life. It
connects the left navigation pane to the right members pane.

When a user expands a node on the left, its children are now loaded on
demand to keep the UI responsive. When a user selects an item, such as a
module or an interface, its members are now fetched and displayed in the
right-hand pane.

(GSoC 2025 - Object Browser: Member View and UI Logic)

Change-Id: I199ebcac20590d1a1ca83ef3ee7759ff34504252
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191923
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>