tdf#148703 fix follow links for recreated page descs
When the creation of a SwPageDesc is undone and then redone,
and that page desc had itself as its own follow, then the
follow link was not being set correctly.

I noticed this when working on tdf#148703, and it is a
small part of that fix that can be merged independently.

Tested with:

CPPUNIT_TEST_NAME="SwCoreUndoTest testPageDescThatFollowsItself" make CppunitTest_sw_core_undo

Change-Id: I798197de78e32694a5819da9025fa84b44e30e98
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191979
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins