tdf#130857 Define widgets in User Data options page in order
Switch the order in which the widgets in the .ui file
are defined so that the order matches the visual appearance
and logical order, which makes sure that tab focus order with the
Qt-based VCL plugins is correct as well when using native Qt
widgets (SAL_VCL_QT_USE_WELDED_WIDGETS=1), where
focus order is (by default) based on widget creation order.

Besides being used in "Tools" -> "Options"
-> "LibreOfficeDev" -> "User Data", this is
also used in a standalone dialog that can be
triggered as follows:

* start Writer
* "Insert" -> "Field" -> "More Fields"
* in the "Document" tab, select "Author" and "Name",
  then "Insert" and close the dialog
* double-click on the newly inserted field to open the
  "Edit Fields" dialog
* click on the "Edit" button

(None of the two use cases are using native Qt
widgets yet. Incorrect focus order was seen
in a WIP branch experimenting with adding support
for the the standalone dialog.)

Change-Id: I0f360f045d3f567f965d24316298bcf4e0194b5b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191959
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>