additions dialog: Move "Show More" button from items to dialog
The "Show More Extensions" button is shown when
PAGE_COUNT (currently 30) extensions are shown,
but there may be more, in order to load/show
the next PAGE_COUNT extensions.

So far, the button was set for each item
in the dialog, and always hidden for all
items except the last one (for which it was
shown when the max was reached).

Instead, move the button to the dialog ancestor,
right below the list of the items.

That's where it semantically belongs, and one button
is enough.

In a quick test, user-visible behavior hasn't
changed in a quick test as follows:

* reduce PAGE_SIZE to 1, so the button shows up
  when there are fewer extensions, rebuild

    diff --git a/cui/source/dialogs/AdditionsDialog.cxx b/cui/source/dialogs/AdditionsDialog.cxx
    index 45a388769527..fa8e39d362ea 100644
    --- a/cui/source/dialogs/AdditionsDialog.cxx
    +++ b/cui/source/dialogs/AdditionsDialog.cxx
    @@ -55,7 +55,7 @@

     #include <bitmaps.hlst>

    -#define PAGE_SIZE 30
    +#define PAGE_SIZE 1

     using namespace css;
     using ::com::sun::star::uno::Reference;

* start Writer
* "Format" -> "Page Style"
* in the "Area" tab, enable "Color"
* click on the "Add color palettes via extension"
  button in the "Colors" section to open the dialog
* click the "Show More Extensions" button a few times

Change-Id: If7c2cad3a2b57e69d84b77d353d769142bb8418e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191971
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins