sc: add a SelectSheetView command to select a sheet view
This adds a command that opens a dialog to show all the available
sheet views for the current sheet. Selecting a sheet view will
switch to that sheet view (or default view) on OK.

This also adds name as a parameter to SheetView class, so we can
store a name of the sheet view. The name is currently generated
and is in a format of "Temp SheetView #".

Change-Id: I11accc30780991f74da97120bb4085ef283a0564
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191978
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>