sc: send sheet view information as part properties via LOKit
This sends the information needed for online to handle show the
sheets are using the sheet view functionality and to show if the
sheet view is synched with the default view or not.

Change getPart to return GetTabNumber instead of CurrentTabForData
or the code will select the wrong sheet when we create or switch to
a sheet view.

Change-Id: I6990af757f60a9a33594e3991dec8b5b3b278b54
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191977
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192147
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>