tdf#130857 qt weld: Evaluate width and height request properties
In QtBuilder evaluate the GtkWidget::width-request [1]
and GtkWidget::height-request [2] properties and
set the corresponding width/height as the minimum
width/height for the QWidget.

This matches what QtInstanceWidget::set_size_request
currently does.

This will e.g. be needed to support the AdditionsDialog/
AdditionsItem, i.e. the dialog to install additional
extensions like color themes.
(Those properties are set for the "imageScreenshot" GtkImage
in cui/uiconfig/ui/additionsfragment.ui.)

Sample way to trigger that dialog (for which support
for using native Qt widgets is not declared yet, but
this commit prepares for that):

* start Writer
* "Format" -> "Page Style"
* in the "Area" tab, enable "Color"
* click on the "Add color palettes via extension"
  button in the "Colors" section to open the dialog

[1] https://docs.gtk.org/gtk3/property.Widget.width-request.html
[2] https://docs.gtk.org/gtk3/property.Widget.height-request.html

Change-Id: Ib4e51e41c0259bcdad4312e927c37103d83a410e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191987
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>