additions dialog: Set proper parent when run from area tab page
Use the new helper introduced in previous commit

    Change-Id: I41c9d9d639bc6e9f523ed2eb74975e2e829836d9
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Oct 7 01:19:03 2025 +0200

        additions dialog: Extract sfx2 helper to run the dialog

to run the additions dialog from the "Area" tab page.

Specify the current dialog as the additions dialog
parent, which ensures that the modal additions dialog
stays on top of its parent dialog as expected for all
of gtk3, gen and qt6, while that was not the case
previously when using the approach of dispatching the
".uno:AdditionsDialog" command.

The dialog can be triggered as follows:

* start Writer
* "Format" -> "Page Style"
* in the "Area" tab, enable "Color"
* click on the "Add color palettes via extension"
  button in the "Colors" section to open the dialog

Similar cases still dispatching the UNO command
exist elsewhere and should be looked into separately.

Change-Id: I480f5275f13c0908372e1a38ef64daca7a27cbfb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191991
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>