additions dialog: Set parent when run to install dicts
This is similar to previous commit

    Change-Id: I480f5275f13c0908372e1a38ef64daca7a27cbfb
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Oct 7 01:29:23 2025 +0200

        additions dialog: Set proper parent when run from area tab page

, but now covers the case where the dialog is run to
install new dictionaries.

Sample scenario:

* start Writer
* "Tools" -> "Spelling"
* press "Options" button
* click on the "Get more dictionaries online..." link

With this commit in place, the dialog now stays
correctly modally on top of the dialog it was
run from.

Also use the helper in SfxApplication::OfaExec_Impl,
which shouldn't cause any change in behavior, but
simplifies and aligns it with the SID_ADDITIONS_DIALOG
case above.

Change-Id: Ib869d9048fd9fc2dbac9768c22e5c3085bcac044
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191993
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>