tdf#166319 sw interdependent redlines: fix redo of direct reject for format
Open the bugdoc, select the format part of the insert-then-format
redline in the manage changes dialog, reject, undo, redo: the result is
<ins>AAA</ins>BBB<ins>CCC</ins> instead of one big insert redline.
Similarly, open the other bugdoc with a delete and a format redline in
it, select the format part of the delete-then-format redline in the
manage changes dialog, reject, undo, redo: the result is
<del>AAA</del><format>BBB</format><del>CCC</del> instead of one big
delete redline.

What happens is that the UI action uses
sw::DocumentRedlineManager::RejectRedlineRange(), while the redo uses
sw::DocumentRedlineManager::RejectRedline(), and while the first
supported direct accepts, the later did not.

Fix the problem by:
1) In sw::DocumentRedlineManager::RejectRedlineRange(), create the undo
   action with the correct depth & directness for direct rejects
2) In DocumentRedlineManager::RejectRedline(), use use
   lcl_RejectOuterFormat() to directly reject a format redline instead
   of the usual lcl_AcceptRejectRedl(), which gives us matching behavior
   for the UI action & the redo.

Writer's interdependent redlines are now in a reasonable state with
this: actions for the redline under cursor work with ins/del when
possible (looking "through" format); while the manage changes dialog
still allows working on the format redline directly if that's wanted.

Change-Id: Idb90dfbd86ac26ff51da8a34731ea75a48748bdf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192015
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>