tdf#168744 editting format of writer input field asserts
Fixing two bugs here. To trigger these bugs, try to do:
  open bugdoc
  right-click -> edit fields
  and then change the selected date format
  and then click OK

First I have to remove an assert added in:
   commit 99055ae98ef1fe67b8db4a8c3167a8acaeaac02f
   Author: Michael Stahl <michael.stahl@allotropia.de>
   Date:   Fri Feb 2 20:10:24 2024 +0100
   tdf#123968 sw: fix assert on importing ooo62823-1.sxw
Which makes no sense, because we use those values all over
the place in  SwSetExpField, and this assert does not
seem to be related to the rest of that bugfix.

Secondly, we have to fix a cast and then a check in
SwFieldPage::InsertField.
SwFieldTypesEnum::Input can, for better or worse, either
refere to an SwInputField or an SwSetExpField, so we
cannot cast to an SwInputField BEFORE we have checked if
it is actually an SwInputField.
Also the condition we were checking here is the wrong way
to determine what kind of subclass this is.

Change-Id: I92c0047a2a3662d3184aad2ec06f565e306d5407
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192067
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins