tdf#165973 sw undo: separate AUTOFORMAT after SPLITNODE
When the user presses RETURN,
they almost certainly want a new paragraph,
but there is much less certainty
that they want the autocorrect.

Since 6.2, these have been a single undo event.
This patch splits it into two undo items,
and places the AUTOFORMAT item last,
so that the first undo restores before the autoformat
and the second undo removes the new paragraph marker.

It was not NECESSARY to affect SwEditShell::SplitNode,
but it seemed prudent to move everything to a higher level,
especially since there was no other time bAutoformat was used.

Change-Id: Ia828e3373608e1e5ecf7792dee2bc04321843d6d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192053
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins