tdf#168406: oox: Restart numbering on 'startAt'
A pptx numbered list can have its numbering restarted at an arbitrary number
with a 'startAt' entry e.g.:
    <a:buAutoNum type="arabicPlain" startAt="3"/>

We already store this value, which ends up in the bullet properties,
but we also need to tell LO to restart numbering using the new value.

In tdf#168406 this saves 8 biking related deaths!

Change-Id: Idb635f6d89f8da35cf91cf0ea266a587f8f46bc6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192161
Tested-by: Jenkins
Reviewed-by: David Gilbert <freedesktop@treblig.org>