sw: text formatting: implement per-line paragraph properties like Word
This doesn't make a whole lot of sense.

Add compatibility setting "HiddenParagraphMarkPerLineProperties" for RTF
and DOCX compatibilityMode < 15.

Apparently what Word's "Compatibility Mode" is doing in case a paragraph
mark has hidden formatting is that it merges the last line of the first
paragraph and the first line of the second paragraph together, and
applies the first paragraph's properties to the line (and the preceding
lines); but for the second line of the second paragraph, it applies the
second paragraph's properties.

This is now implemented here; firstly, by adding a flag to the
MergedPara's Extents so that the situation can be distinguished (if the
paragraphs are joined by a delete redline, Word does something different
of course).  Because it's possible that the hidden paragraph break is on
a paragraph that doesn't have any extents, but a preceding paragraph has
extents that are affected, this sometimes requires 0-length dummy
extents.

FindParaPropsNodeIgnoreHidden() sets the last paragraph as
pParaPropsNode, which is used for all non-per-line properties.

Note that it's somewhat likely that the various Update functions in
txtfrm.cxx don't maintain the isHiddenParaMerge flag on extents
correctly, so it may look different than Word when editing.

Currently it affects these properties:

* line spacing

* tab stops

These are now set per line by SwTextIter::Next() calling
SwAttrIter::GetTextNodeForLinePropsWordCompat() and a factored out
SwLineInfo::InitLineInfo().

Evidently Word also does adjustment this way, but it's not implemented
here.

Conflicts:
    sw/source/core/text/inftxt.cxx

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192077
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>
(cherry picked from commit 0849ddd0b1b3c384c5f3de8fe0bbb9df558fa786)

sw: fix too early reset of m_oParagraphBreak
Thanks to Mike Kaganski for finding this
(regression from commit 0849ddd0b1b3c384c5f3de8fe0bbb9df558fa786)

Change-Id: I216d9e2afdac9ab6f97c0ea822d4d501689df7a6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192102
Reviewed-by: Thorsten Behrens <thorsten.behrens@collabora.com>
Tested-by: allotropia jenkins <jenkins@allotropia.de>