tdf#164641 sc: no SpellPopup when IsCellMarked
Just like Writer (forever) and Draw/Impress (26.2),
don't show the spelling context menu on a right-click
if the cell is selected (unless it is the cell being
actively edited somehow).

The problem is that the spellPopup menu is NOT
based on the cell under the mouse,
but on the Point of the selected area (bPosIsInEditView).
This has always been super confusing and unexpected,
so simply refuse to handle spell-checking inside a selection.

Note that even in selection, you can click into the
"Input Line" to put the cell into bPosIsInEditView.
In that case do still allow a spellcheck context menu
even though it is technically still in a selection.

Change-Id: I9a13447614d35c43f9aa01dd9a9ecf720d0260a1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192084
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>