Related: tdf#80224 A blind attempt to fix conditions
... introduced in commit 43679f94b45f4d9e120c64a3fb5cc3ee77f12b11
(Fix tdf#80224 Custom text color changed to black on .PPTX export,
2015-08-25).

Those conditions, presumably, intended to use GetProperty in the
non-bCheckDirect case. But the implementation effectively made the
bCheckDirect case also call GetProperty, when state check fails,
which, in the end, has the same effect, as calling GetProperty in
all cases.

I attempted to fix it like this:

    bCheckDirect ? GetDirectProperty(rXPropSet, rXPropState, u"Foo"_ustr)
                 : GetProperty(rXPropSet, u"Foo"_ustr)

but it failed CppunitTest_chart2_export3 then:

    Test name: testFormattedChartTitles::TestBody
    equality assertion failed
    - Expected: 1
    - Actual  : 0
    - In <>, XPath '/c:chartSpace/c:chart/c:title/c:tx/c:rich/a:p[1]/a:r[5]/a:rPr/a:uFillTx' number of nodes is incorrect

So this just simplifies the calls to what they effectively did all
this time.

There is no failing document; I found this accidentally, reading
the code.

Change-Id: I6581c23e395a74140daa30a06cdca27b57f4e6a6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192111
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>