revert some changes to SwFieldPage::InsertField
where I had no idea of the mess I was stepping into.
And add some comments for anyone else looking at this.

This reverts parts of
    commit da024246e6ead07f678ab17e42fdb9b071a65c1a
    Author: Noel Grandin <noelgrandin@gmail.com>
    Date:   Sun Jul 6 13:10:03 2025 +0200
    convert SwInputFieldSubType to scoped enum
and
    commit 2d7883aba225305f7d9acd213c7e21c0012e6c0c
    Author: Noel Grandin <noelgrandin@gmail.com>
    Date:   Tue Oct 7 16:11:37 2025 +0200
    tdf#168744 editting format of writer input field asserts

i.e. I am returning this specific chunk of logic to the effective
state it was in before those two commits.
I say effective, because the original code was using the "wrong"
define constant in the if statement, so I used the constant that
would result in the same comparison.

Change-Id: I5164e72cbbf823fee4278e0fad2ad7216be47882
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192095
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>