tdf#168116: an ugly hack to handle linebreak-before-hidden-paragraph-mark
Word eliminates a completely empty line ending with a hidden paragraph
mark, which may appear when a paragraph has a line break followed by such
a break, even when the following is not another paragraph, but a table.
This change detects this event, and removes the linebreak instead.

Known problems:

1. If there is a completely hidden paragraph between the paragraph with
trailing line break and hidden mark, this doesn't work. E.g., this RTF
markup:

AAAA\line
{\v\par}
{\v\par}
\intbl BBBB\cell \trowd \lastrow \clbrdrt\brdrs\brdrw10 \clbrdrl\brdrs\brdrw10 \clbrdrb\brdrs\brdrw10 \clbrdrr\brdrs\brdrw10 \cellx5000\row \pard

Still, this change allows to fix the bugdoc layout.

2. When formatting marks are shown, the paragraph ends with pilcrow, not
a linebreak symbol. This is minor.

Change-Id: I66b06cdac9ab41bcccffab669f7caffe0a23b686
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192104
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>