tdf#168559 PPTX imp: fix missing custom level list style for outline shapes
Export the bugdoc to PPTX, reopen it in Impress, go to the end of the
first bullet, enter, tab, indent grows to a large value, while it should
only grow to match the existing other already indented bullet.

This happens because the shape text's paragraphs have a numbering rule
defined for the current list level, but the other list levels are left
untouched by the PPTX import.

Fix the problem by optionally giving the shape's list style to
TextParagraphProperties::pushToPropSet(), and if that's given, then use
the list style to initialize the numbering rules of the paragraph for
all non-current levels. (Current level was initialized already.) Not all
possible properties are set, but left/first margin is done, which is the
primary use-case of the bug.

This fixes, the bug, but a full PPTX export+import+export+import cycle
is still bad, probably because something is not imported on the
masterpage, that's still to be done. (PPTX export+import+export then
edit in PowerPoint works.)

Change-Id: I3575b986faaacb4103e6dde083b13b10b7d957c1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192168
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins