Revert "Prevent cursor invalidation if the cursor position doesn't change."
This reverts commit a260cc52b2fae1382805b4389c95f29ed8671f42.
This commit causes the view jump on writer.

This commit originally merged for the following issue

Open two writer documents side by side includes hyperlink
Click the hyperilnk one of them and see hyperlink popup
Type into the other document

There was a glich on hyperlink popup in the first document

Issue doesn't exist even without this patch. So we remove it completely.

Change-Id: I8a341d66a963ed1943d9f666ba026d7eca240f08
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192157
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>