tdf#51850 sw: AutoCaption for a single pasted image
AutoCaption mostly isn't enabled for clipboard paste, except for one
special case.

For the case when the source is outside LO, let
SwTransferable::PasteGrf() set a flag that the caption should be
inserted (the SwPasteSdr::Insert case), which can then be handled at the
end.

PasteFileList() appears to be used when an image file is copied from
Nautilus.

For internal paste, it gets more complicated because SwFEShell::Paste()
knows what was inserted, but caption can only be added later, outside of
any StartAction()/EndAction(), because if an action is pending,
SwWrtShell::GetSelectionType() does not return useful results, and the
lcl_InsertLabel() ends up erroneously inserting a text node next to the
SwGrfNode.

Change-Id: I3e16f407fd409a0f315be139376ad9e5f0b526cd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192165
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>